open_ai_api/requirements.txt
Flask==2.3.2
openai==1.0.0
python-dotenv==0.21.1
requests==2.28.2
